% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoMcl.R
\name{clustered}
\alias{clustered}
\title{multiple mediators clustered data}
\usage{
clustered(
  data,
  Sname,
  Wnames = NULL,
  Xnames,
  Aname,
  Mnames,
  Mfamily = "binomial",
  Yname,
  Yfamily = "binomial",
  cluster_opt_a = "FE.glm",
  cluster_opt_m = "FE.glm",
  cluster_opt_y = "FE.glm",
  interaction_fitm2 = c("AM.M1"),
  interaction_fity = c("AM.M1", "AM.M2", "Mint", "AMint"),
  num_folds = 1,
  learners = c("SL.glm"),
  learners_a = c("SL.glm"),
  learners_m = c("SL.glm"),
  learners_y = c("SL.glm")
)
}
\description{
multiple mediators clustered data
}
